Move controller logic from user routes into user-controller.js
